+++ prisma-crm/prisma-crm-ejb/src/main/java/Entities/VehiculeMaintenance.java
@@ -18,144 +18,80 @@ public class VehiculeMaintenance implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private int id;
 	@ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
 	private Vehicule vehicule;
-	
+
 	private ServiceType serviceType;
 	private RepairStatus repairStatus;
 	private float totalprice;
-	
+
 	private Date maintainceDate;
-	
-	private float odometer;
-	
-	
-	
-	
-	
+
+	private int odometer;
+
 	public int getId() {
 		return id;
 	}
-
-
-
-
 
 	public void setId(int id) {
 		this.id = id;
 	}
 
-
-
-
-
 	public Vehicule getVehicule() {
 		return vehicule;
 	}
-
-
-
-
 
 	public void setVehicule(Vehicule vehicule) {
 		this.vehicule = vehicule;
 	}
 
-
-
-
-
 	public ServiceType getServiceType() {
 		return serviceType;
 	}
-
-
-
-
 
 	public void setServiceType(ServiceType serviceType) {
 		this.serviceType = serviceType;
 	}
 
-
-
-
-
 	public float getTotalprice() {
 		return totalprice;
 	}
-
-
-
-
 
 	public void setTotalprice(float totalprice) {
 		this.totalprice = totalprice;
 	}
 
-
-
-
-
 	public Date getMaintainceDate() {
 		return maintainceDate;
 	}
-
-
-
-
 
 	public RepairStatus getRepairStatus() {
 		return repairStatus;
 	}
 
-
-
-
-
 	public void setRepairStatus(RepairStatus repairStatus) {
 		this.repairStatus = repairStatus;
 	}
-
-
-
-
 
 	public void setMaintainceDate(Date maintainceDate) {
 		this.maintainceDate = maintainceDate;
 	}
 
-
-
-
-
-	public float getOdometer() {
+	public int getOdometer() {
 		return odometer;
 	}
 
-
-
-
-
-	public void setOdometer(float odometer) {
+	public void setOdometer(int odometer) {
 		this.odometer = odometer;
 	}
-
-
-
-
 
 	public static long getSerialversionuid() {
 		return serialVersionUID;
 	}
 
-
-
-
-
-	public VehiculeMaintenance(Vehicule vehicule, ServiceType serviceType, float totalprice, float odometer) {
+	public VehiculeMaintenance(Vehicule vehicule, ServiceType serviceType, float totalprice, int odometer) {
 		super();
 		this.vehicule = vehicule;
 		this.serviceType = serviceType;
@@ -163,27 +99,15 @@ public class VehiculeMaintenance implements Serializable {
 		this.odometer = odometer;
 	}
 
-
-
-
-
-	public VehiculeMaintenance(Vehicule vehicule, float totalprice, float odometer) {
+	public VehiculeMaintenance(Vehicule vehicule, float totalprice, int odometer) {
 		super();
 		this.vehicule = vehicule;
 		this.totalprice = totalprice;
 		this.odometer = odometer;
 	}
 
-
-
-
-
 	public VehiculeMaintenance() {
 	}
-
-
-
-
 
 	@Override
 	public String toString() {
@@ -191,5 +115,5 @@ public class VehiculeMaintenance implements Serializable {
 				+ ", repairStatus=" + repairStatus + ", totalprice=" + totalprice + ", maintainceDate=" + maintainceDate
 				+ ", odometer=" + odometer + "]";
 	}
-   
+
 }
